requirements.txt
praw==7.7.1
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
feedparser==6.0.10
selenium==4.16.0
webdriver-manager==4.0.1 